let exception propagate by default
Hiding the exception by default might cause less noise in asterisk when
there is an error, but then it's harder to debug. The right solution
would be to add a "silent" flag to xivo-confgen, and use the flag in all
the asterisk configuration files, i.e. replace "xivo-confgen
asterisk/sip.conf" with "xivo-confgen -s asterisk/sip.conf", for example.